\name{MxVAR-package}
\alias{MxVAR-package}
\alias{MxVAR}
\docType{package}
\title{
\packageTitle{MxVAR}
}
\description{
\packageDescription{MxVAR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MxVAR}
\packageIndices{MxVAR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MxVAR}

Maintainer: \packageMaintainer{MxVAR}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
